phase-1-ingestion/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "second-brain-ingestion"
version = "0.1.0"
description = "Second Brain ingestion pipeline: watch vault, chunk, embed, index in Qdrant."
requires-python = ">=3.12"
dependencies = [
    "watchdog>=4.0.0",
    "voyageai>=0.3.0",
    "qdrant-client[fastembed]>=1.12.0",
    "python-frontmatter>=1.1.0",
    "mistune>=3.0.2",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.1.0",
    "tiktoken>=0.7.0",
    "tenacity>=8.3.0",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
ingestion = "ingestion.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
